add user receipt email + address PR review fixes
- use GetEnvVarOrFail for envs (incl. CONTACT_REQUIRE_CAPTCHA); types moved to src/types/nodemailer.d.ts
- Adjusted rate limit to 5 requests per minute per IP
- Added confirmation email to submitter
